// types.ts
export type Role = 'admin' | 'worker' | 'driver';
export type Language = 'fr' | 'ar';
// ...
export interface WorkerTransaction {
  id: string;
  type: 'payment' | 'advance' | 'absence';
  amount: number;
  date: string;
  note?: string;
}

export interface Worker {
  id: string;
  fullName: string;
  birthday: string;
  phone: string;
  email?: string;
  address: string;
  idCardNumber: string;
  photo?: string;
  role: Role;
  paymentType: 'day' | 'month';
  amount: number; // This is the salary (per day or month)
  username: string;
  password?: string;
  absences: number; // Count of days
  totalPaid: number;
  history: WorkerTransaction[];
}